pyproject.toml
[project]
name = "code-wise"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "openai ==1.72.0",
    "python-dotenv ==1.0.1",
    "requests ==2.32.3",
    "pyside6 ==6.9.0",
    "cognitive-complexity ==1.3.0",
]

[tool.uv]
dev-dependencies = [
    "iniconfig ==2.0.0",
    "packaging ==24.2",
    "pluggy ==1.5.0",
    "pytest ==8.3.4",
    "pytest-cov ==6.0.0",
    "ruff ==0.11.4",
    "isort ==6.0.1",
    "black ==25.1.0",
    "pre-commit ==4.2.0"
]

[tool.setuptools]
packages = ["source"]
package-dir = {"source" = "source"}

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[tool.black]
line-length = 120
target-version = ["py313"]
skip-string-normalization = true
